requirements.txt
django-eav2>=0.13.0,<1.0.0
djangorestframework>=3.11.0,<4.0.0
Markdown>=3.2.2,<4.0.0
django-guardian>=2.3.0,<3.0.0
PyYAML>=5.3.1,<6.0.0
uritemplate>=3.0.1,<4.0.0
Faker==4.0.2
django-filter>=2.3.0,<3.0.0
djangorestframework-guardian>=0.3.0,<1.0.0
Pillow>=7.2.0,<8.0.0
django-extensions>=3.0.5,<4.0.0
psycopg2-binary>=2.8,<3.0
django-environ>=0.4.5,<1.0.0
django-cors-headers>=3.5.0,<4.0.0
django-crispy-forms>=1.9.2,<2.0.0
drf-nested-routers>=0.91,<1.0
djangorestframework-csv>=2.1.0,<3.0.0
inflection>=0.5.1,<1.0.0
ipython>=7.17.0,<8.0.0
django-silk>=4.1.0,<5.0.0
drf-tweaks>=0.9.5,<1.0.0
djoser>=2.0.5,<3.0.0
mailchimp3>=3.0.14,<4.0.0